Remove railtie, speed up tests, clean up test structure.
Remove railtie which should go in `prx_auth-rails` gem. That gem can include
controller methods, automatic middleware insertion, and a dependency on this
gem.

Clean up test directory structure and guardfile.

Go from from 1.7 assertions/s (post railtie-related test removal) to
104.4 assertions/s.
