+++ Rakefile
@@ -3,7 +3,8 @@ require 'rake'
 require 'rake/testtask'
 
 Rake::TestTask.new do |t|
-  t.pattern = 'test/*test.rb'
+  t.libs << 'test'
+  t.pattern = 'test/**/*test.rb'
 end
 
 task default: :test
