--- lib/rack/prx_auth/controller_methods.rb
@@ -1,11 +1,0 @@
-class Rack::PrxAuth
-  module ControllerMethods
-    def prx_auth_token
-      request.env['prx.auth']
-    end
-
-    def prx_authenticated?
-      !!prx_auth_token
-    end
-  end
-end
