+++ lib/rack/prx_auth/version.rb
@@ -1,5 +1,5 @@
 module Rack
   class PrxAuth
-    VERSION = "0.0.5"
+    VERSION = "0.0.6"
   end
 end
