+++ test/rack/prx_auth_test.rb
@@ -1,4 +1,4 @@
-require_relative 'minitest_helper'
+require 'test_helper'
 
 describe Rack::PrxAuth do
   let(:app) { Proc.new {|env| env } }
@@ -43,11 +43,13 @@ describe Rack::PrxAuth do
     end
 
     it 'attaches claims to request params if verification passes' do
-      JSON::JWT.stub(:decode, claims) do
-        prxauth.call(env)['prx.auth'].tap do |token|
-          token.must_be_instance_of Rack::PrxAuth::TokenData
-          token.attributes.must_equal claims
-          token.user_id.must_equal claims['sub']
+      prxauth.stub(:decode_token, claims) do
+        prxauth.stub(:valid?, true) do
+          prxauth.call(env)['prx.auth'].tap do |token|
+            token.must_be_instance_of Rack::PrxAuth::TokenData
+            token.attributes.must_equal claims
+            token.user_id.must_equal claims['sub']
+          end
         end
       end
     end
@@ -71,60 +73,6 @@ describe Rack::PrxAuth do
       Rack::PrxAuth::Certificate.stub(:new, Proc.new{|l| loc = l}) do
         Rack::PrxAuth.new(app, cert_location: :location)
         loc.must_equal :location
-      end
-    end
-  end
-
-  describe Rack::PrxAuth::Certificate do
-    let(:subject) { Rack::PrxAuth::Certificate.new }
-    let(:certificate) { subject }
-
-    describe '#initialize' do
-      it 'allows setting the location of the certificates' do
-        cert = Rack::PrxAuth::Certificate.new('http://example.com')
-        cert.cert_location.must_equal URI('http://example.com')
-      end
-
-      it 'defaults to DEFAULT_CERT_LOC' do
-        cert = Rack::PrxAuth::Certificate.new
-        cert.cert_location.must_equal Rack::PrxAuth::Certificate::DEFAULT_CERT_LOC
-      end
-    end
-
-    describe '#valid?' do
-      it 'validates the token with the public key' do
-        token, key = nil, nil
-        certificate.stub(:public_key, :public_key) do
-          JSON::JWT.stub(:decode, Proc.new {|t, k| token, key = t, k }) do
-            certificate.valid?(:token)
-          end
-        end
-
-        token.must_equal :token
-        key.must_equal :public_key
-      end
-
-      it 'returns false if verification fails' do
-        JSON::JWT.stub(:decode, Proc.new {|t, k|
-          raise JSON::JWT::VerificationFailed }) do
-          certificate.wont_be :valid?, :token
-        end
-      end
-
-      it 'returns true if verification passes' do
-        JSON::JWT.stub(:decode, {}) do
-          certificate.must_be :valid?, :token
-        end
-      end
-    end
-
-    describe '#certificate' do
-      it 'calls fetch if unprimed' do
-        def certificate.fetch
-          :sigil
-        end
-
-        certificate.send(:certificate).must_equal :sigil
       end
     end
   end
