+++ test/test_helper.rb
@@ -7,5 +7,3 @@ require 'rack/prx_auth'
 require 'minitest/autorun'
 require 'minitest/spec'
 require 'minitest/pride'
-require 'minitest/stub_any_instance'
-require 'lumberjack' rescue nil
